feat:SC-397 Add createSightTest mutation and support to MsSQL
This PR adds support to use MySQL and MsSQL at the same time on clinical service and also creates a new mutation `createSightTest `

For now I'm using a pre-release version of common jar. Update to a release version will be done the the PR on common is merged

NOTE: **We MUST merge this PR at the same time PR below**, since we are adding a SQL Server to the service.
[https://dev.azure.com/Specsavers/Socrates-Cloud/_git/terraform-socrates_cloud/pullrequest/16628]()